refactoring move proposers
