+++ src/mcmm.py
@@ -21,7 +21,6 @@ runs MMFF + accept/reject.
 
 import math
 
-import ase
 import numpy as np
 import torch
 from rdkit import Chem
@@ -229,111 +228,9 @@ def _ordered_backbone_residues(mol: Chem.Mol) -> list[tuple[int, int, int]]:
     return best_cycle
 
 
-# ---------------------------------------------------------------------------
-# Side-chain enumeration — used by the MCMM proposer for full-mol moves
-# ---------------------------------------------------------------------------
-
-
-def _backbone_atom_set(mol: Chem.Mol) -> set:
-    """
-    Return the set of atom indices that lie on the macrocycle backbone.
-
-    Backbone atoms are the atoms on the largest ring as identified by
-    RDKit ring perception (see `_ordered_macrocycle_atoms`). Used as
-    the "stop set" for the side-chain BFS in `_side_chain_group` —
-    atoms outside this set are side-chain candidates; atoms inside
-    are part of the ring and must not be crossed during the BFS.
-
-    Params:
-        mol: Chem.Mol : a cyclic molecule with a macrocycle ring
-    Returns:
-        set[int] : ring atom indices; empty if no qualifying ring exists
-    """
-    return set(_ordered_macrocycle_atoms(mol))
-
-
-def _side_chain_group(mol: Chem.Mol, atom_idx: int, backbone_atom_set: set) -> set:
-    """
-    BFS from `atom_idx` through non-backbone bonds.
-
-    Returns the set of atom indices reachable from `atom_idx` without
-    crossing any backbone atom (the starting atom itself is excluded
-    from the result; only its non-backbone neighbours and beyond are
-    included). For a backbone atom in a cyclic peptide, this returns
-    the side chain attached at that residue — Hα, Cβ, side-chain
-    branches, etc. — without leaking into adjacent residues' side
-    chains via the macrocycle.
-
-    For a non-backbone starting atom, the result is the connected
-    non-backbone component containing it (minus the starting atom),
-    which is rarely what callers want; pass a backbone atom.
-
-    Params:
-        mol: Chem.Mol : input molecule
-        atom_idx: int : starting atom (typically a backbone atom)
-        backbone_atom_set: set[int] : atoms forming the macrocycle ring;
-            traversal does not cross these
-    Returns:
-        set[int] : reachable atom indices, excluding `atom_idx` itself
-    """
-    side_chain: set = set()
-    queue: list = []
-    for nb in mol.GetAtomWithIdx(atom_idx).GetNeighbors():
-        nb_idx = nb.GetIdx()
-        if nb_idx not in backbone_atom_set:
-            queue.append(nb_idx)
-    while queue:
-        idx = queue.pop()
-        if idx in side_chain:
-            continue
-        side_chain.add(idx)
-        for nb in mol.GetAtomWithIdx(idx).GetNeighbors():
-            nb_idx = nb.GetIdx()
-            if nb_idx in backbone_atom_set or nb_idx in side_chain:
-                continue
-            queue.append(nb_idx)
-    return side_chain
-
-
-def _compute_window_downstream_sets(
-    mol: Chem.Mol, window: tuple, backbone_atom_set: set
-) -> list:
-    """
-    Compute the full-mol atom indices that should rotate per dihedral
-    when a DBT move acts on `window`.
-
-    For dihedral k around bond (window[k+1], window[k+2]):
-      - Window backbone atoms strictly downstream: window[k+3..6].
-      - Side chains of window[k+2..6] (the pivot atom k+2's side chain
-        rotates with the local frame at the pivot; downstream backbone
-        atoms' side chains rigidly follow their parents).
-
-    The pivot atom window[k+2] itself stays on the rotation axis and
-    does not move. Side chains of window[0..k+1] are upstream of the
-    bond and do not rotate.
-
-    Params:
-        mol: Chem.Mol : input molecule
-        window: tuple[int, ...] : 7 atom indices, in chain order
-        backbone_atom_set: set[int] : from `_backbone_atom_set(mol)`,
-            passed in to avoid recomputation across windows
-    Returns:
-        list of 4 frozenset[int] : per-dihedral rotation set, suitable
-            for `concerted_rotation.apply_dihedral_changes_full_mol`
-    """
-    side_chains = {
-        atom_idx: _side_chain_group(mol, atom_idx, backbone_atom_set)
-        for atom_idx in window
-    }
-    downstream_sets: list = []
-    for k in range(4):
-        rotated: set = set()
-        rotated.update(side_chains[window[k + 2]])
-        for j in range(k + 3, 7):
-            rotated.add(window[j])
-            rotated.update(side_chains[window[j]])
-        downstream_sets.append(frozenset(rotated))
-    return downstream_sets
+# Side-chain enumeration helpers (`_backbone_atom_set`, `_side_chain_group`,
+# `_compute_window_downstream_sets`) moved to `src/proposers.py` — they are
+# now consumed exclusively by the proposer factories that also live there.
 
 
 # ---------------------------------------------------------------------------
@@ -1522,468 +1419,18 @@ class ReplicaExchangeMCMMDriver:
 
 
 # ---------------------------------------------------------------------------
-# Real-mol proposer factory
+# Back-compat re-exports — proposer factories live in src/proposers.py.
+# Imports placed at the bottom of mcmm.py so the helpers consumed by
+# proposers (`_ordered_macrocycle_atoms`, `enumerate_backbone_windows`) are
+# defined before proposers.py loads, preventing a partial-import error.
 # ---------------------------------------------------------------------------
 
 
-def make_mcmm_proposer(
-    mol: Chem.Mol,
-    hardware_opts,
-    calc,
-    drive_sigma_rad: float = 0.1,
-    closure_tol: float = 0.01,
-    score_chunk_size: int = 500,
-    mmff_backend: str = "gpu",
-    seed: int = 0,
-):
-    """
-    Build a `batch_propose_fn` for `ReplicaExchangeMCMMDriver` that
-    proposes DBT moves on the backbone windows of `mol`, batches MMFF +
-    MACE across walkers per call, and returns per-walker
-    `(new_coords, new_energy, det_j, success)` tuples.
-
-    Per-call pipeline:
-
-      1. **Per-walker move generation** (CPU, sequential): pick a random
-         backbone window, drive dihedral, and `drive_delta ~ N(0,
-         drive_sigma_rad²)`. Run `concerted_rotation.propose_move` on
-         the 7-atom backbone window positions to solve for the closure
-         deltas. Walkers whose closure fails are flagged.
-      2. **Full-mol coordinate update** (CPU, sequential): for
-         successful walkers, replay the per-dihedral rotations on the
-         full atom array via
-         `concerted_rotation.apply_dihedral_changes_full_mol`,
-         transporting side-chain atoms rigidly with their backbone
-         parents according to the precomputed
-         `_compute_window_downstream_sets` for the chosen window.
-      3. **Batched MMFF** (GPU, one call): stage every successful
-         candidate as a conformer on a shared throwaway mol and run
-         `nvmolkit.mmffOptimization.MMFFOptimizeMoleculesConfs`
-         (`mmff_backend='gpu'`) or RDKit's serial MMFF
-         (`mmff_backend='cpu'`) for in-place minimisation.
-      4. **Batched MACE** (GPU, chunked): score every minimised
-         candidate via `_mace_batch_energies` in chunks of
-         `score_chunk_size`.
-      5. **Return** `(coords_tensor, energy_float, det_j_float,
-         success_bool)` per walker, in walker order. Failed walkers
-         pass through with their pre-move coords and `success=False`
-         so the driver's `apply_proposal` rejects without further work.
-
-    Topology — backbone windows, side-chain groups, the throwaway-mol
-    template — is captured at factory build time and reused per call,
-    keeping the per-step CPU overhead bounded by the move-generation
-    loop and the conformer-staging step.
-
-    Lazy import of `_mace_batch_energies` from `confsweeper` avoids the
-    confsweeper → mcmm circular dependency at module load time.
-
-    Params:
-        mol: Chem.Mol : a head-to-tail cyclic peptide with explicit Hs.
-            Topology is captured at factory-build time; the mol must
-            not be mutated structurally afterwards (conformer additions
-            and edits are fine).
-        hardware_opts : nvmolkit hardware options for batched MMFF
-            (only consulted when `mmff_backend='gpu'`).
-        calc : MACECalculator from `get_mace_calc()`.
-        drive_sigma_rad: float : Gaussian standard deviation for the
-            drive-angle perturbation in radians (default 0.1 ≈ 5.7°).
-            Larger values give bigger moves at lower closure-success
-            rate; couples to closure_tol per docs/mcmm_plan.md.
-        closure_tol: float : passed through to `propose_move` as the
-            maximum r5+r6 displacement-norm tolerated as ring-closed.
-        score_chunk_size: int : MACE per-batch forward pass cap
-            (default 500, matches `_minimize_score_filter_dedup`).
-        mmff_backend: str : 'gpu' (nvmolkit batched CUDA, default) or
-            'cpu' (RDKit serial). MMFF runs on the throwaway mol.
-        seed: int : base seed for the move-RNG; deterministic across
-            replicate runs.
-    Returns:
-        callable : `batch_propose_fn(coords_list) -> list[tuple]` matching
-            the contract expected by `ParallelMCMMDriver` and
-            `ReplicaExchangeMCMMDriver`.
-    Raises:
-        ValueError: if `mol` has no enumerable backbone windows (input
-            is not a cyclic peptide of ≥ 3 residues).
-    """
-    from concerted_rotation import (
-        N_DIHEDRALS,
-        apply_dihedral_changes_full_mol,
-        propose_move,
-    )
-
-    if mmff_backend not in ("gpu", "cpu"):
-        raise ValueError(
-            f"unknown mmff_backend {mmff_backend!r}; expected 'gpu' or 'cpu'"
-        )
-
-    windows = enumerate_backbone_windows(mol)
-    if not windows:
-        raise ValueError(
-            "mol has no enumerable backbone windows; "
-            "check that it is a head-to-tail cyclic peptide of at least 3 residues"
-        )
-    backbone_atoms = _backbone_atom_set(mol)
-    window_downstream_sets = [
-        _compute_window_downstream_sets(mol, w, backbone_atoms) for w in windows
-    ]
-    n_windows = len(windows)
-    n_atoms = mol.GetNumAtoms()
-    atomic_nums = [a.GetAtomicNum() for a in mol.GetAtoms()]
-
-    # Throwaway-mol template: structure-only, no conformers. Cloning per
-    # call is required because nvmolkit MMFF mutates conformers in place
-    # and we don't want to corrupt walker state across step() calls.
-    template_mol = Chem.Mol(mol)
-    template_mol.RemoveAllConformers()
-
-    rng = np.random.default_rng(seed)
-
-    # Per-call cumulative diagnostic counters. Attached to the returned
-    # proposer function as `.stats` so callers (`get_mol_PE_mcmm`, tests)
-    # can read them after the run to diagnose acceptance regressions —
-    # especially the "1 basin" pathology on small peptides where DBT
-    # closure fails on most moves.
-    stats = {
-        "n_proposed": 0,
-        "n_closure_failures": 0,
-        "n_closure_successes": 0,
-    }
-
-    def batch_propose_fn(coords_list):
-        # Lazy import: confsweeper imports from mcmm at module load time;
-        # importing _mace_batch_energies here defers resolution until the
-        # closure is actually called, breaking the circular dependency.
-        from confsweeper import _mace_batch_energies
-
-        n_walkers = len(coords_list)
-        stats["n_proposed"] += n_walkers
-
-        # Stage 1: per-walker DBT closure on the backbone window.
-        # Successful walkers contribute a (new_full_coords, det_j) entry.
-        successful_meta: list = []
-        success_walker_indices: list = []
-        for w_idx, coords in enumerate(coords_list):
-            window_idx = int(rng.integers(n_windows))
-            window = windows[window_idx]
-            drive_idx = int(rng.integers(N_DIHEDRALS))
-            drive_delta = float(rng.normal(0.0, drive_sigma_rad))
-
-            coords_np = coords.detach().cpu().numpy().astype(np.float64)
-            window_pos = coords_np[list(window)]
-            result = propose_move(
-                window_pos, drive_idx, drive_delta, closure_tol=closure_tol
-            )
-            if not result.success:
-                continue
-            new_full = apply_dihedral_changes_full_mol(
-                coords_np,
-                list(window),
-                result.deltas,
-                window_downstream_sets[window_idx],
-            )
-            successful_meta.append(
-                {"new_full": new_full, "det_j": float(result.det_jacobian)}
-            )
-            success_walker_indices.append(w_idx)
-
-        stats["n_closure_successes"] += len(successful_meta)
-        stats["n_closure_failures"] += n_walkers - len(successful_meta)
-
-        # Stage 2: short-circuit if every walker failed.
-        if not successful_meta:
-            return [(coords_list[i], 0.0, 0.0, False) for i in range(n_walkers)]
-
-        # Stage 3: stage successful candidates as conformers on a fresh
-        # throwaway mol, run batched MMFF.
-        throwaway = Chem.Mol(template_mol)
-        for meta in successful_meta:
-            conf = Chem.Conformer(n_atoms)
-            new_full = meta["new_full"]
-            for a_idx in range(n_atoms):
-                x, y, z = new_full[a_idx]
-                conf.SetAtomPosition(a_idx, (float(x), float(y), float(z)))
-            throwaway.AddConformer(conf, assignId=True)
-
-        if mmff_backend == "gpu":
-            from nvmolkit.mmffOptimization import MMFFOptimizeMoleculesConfs
-
-            MMFFOptimizeMoleculesConfs([throwaway], hardwareOptions=hardware_opts)
-        else:
-            from rdkit.Chem import AllChem as _AllChem
-
-            for cid in [c.GetId() for c in throwaway.GetConformers()]:
-                _AllChem.MMFFOptimizeMolecule(throwaway, confId=cid)
-
-        # Stage 4: batched MACE scoring, chunked.
-        post_mmff_conf_ids = [c.GetId() for c in throwaway.GetConformers()]
-        energies: list = []
-        for start in range(0, len(post_mmff_conf_ids), score_chunk_size):
-            chunk_ids = post_mmff_conf_ids[start : start + score_chunk_size]
-            ase_mols = [
-                ase.Atoms(
-                    positions=throwaway.GetConformer(cid).GetPositions(),
-                    numbers=atomic_nums,
-                )
-                for cid in chunk_ids
-            ]
-            energies.extend(_mace_batch_energies(calc, ase_mols))
-
-        # Stage 5: assemble per-walker proposals in walker order.
-        proposals: list = [None] * n_walkers
-        for slot, w_idx in enumerate(success_walker_indices):
-            cid = post_mmff_conf_ids[slot]
-            new_coords = torch.tensor(
-                throwaway.GetConformer(cid).GetPositions(), dtype=torch.float64
-            )
-            proposals[w_idx] = (
-                new_coords,
-                float(energies[slot]),
-                successful_meta[slot]["det_j"],
-                True,
-            )
-        # Failed walkers get a no-op proposal in their original slot.
-        for w_idx in range(n_walkers):
-            if proposals[w_idx] is None:
-                proposals[w_idx] = (coords_list[w_idx], 0.0, 0.0, False)
-
-        return proposals
-
-    # Expose cumulative stats so the orchestrator (get_mol_PE_mcmm) can
-    # read closure-failure rates and diagnose "1 basin" regressions.
-    batch_propose_fn.stats = stats
-    return batch_propose_fn
-
-
-# ---------------------------------------------------------------------------
-# Cartesian-kick proposer — GOAT-style topology-preserving move
-# ---------------------------------------------------------------------------
-
-
-def make_cartesian_kick_proposer(
-    mol: Chem.Mol,
-    hardware_opts,
-    calc,
-    sigma_kick_a: float = 0.1,
-    score_chunk_size: int = 500,
-    mmff_backend: str = "gpu",
-    seed: int = 0,
-):
-    """
-    Build a `batch_propose_fn` that applies an isotropic Gaussian kick
-    to all atom positions, MMFF-relaxes, and MACE-scores. The
-    GOAT-flavour move type: complements DBT's dihedral-space
-    parameterisation by reaching geometries that small dihedral
-    perturbations can't (side-chain rotamer flips, depsipeptide ester
-    rearrangements, etc.).
-
-    Per-call pipeline:
-
-      1. **Per-walker kick** (CPU, vectorised): add `N(0, sigma_kick_a²)`
-         noise independently to every atom-coordinate of every walker.
-         No closure step or backbone parameterisation; rely on MMFF to
-         pull bond lengths and ring sp² angles back to equilibrium.
-      2. **Batched MMFF** (GPU, one call): stage every kicked candidate
-         as a conformer on a shared throwaway mol and run
-         `nvmolkit.mmffOptimization.MMFFOptimizeMoleculesConfs`.
-      3. **Batched MACE** (GPU, chunked) via `_mace_batch_energies`.
-      4. **Return** `(coords_tensor, energy_float, det_j=1.0,
-         success=True)` per walker. The Gaussian kick is symmetric in
-         coordinate space so detailed balance needs no Jacobian
-         correction; the post-MMFF state is treated as the proposal
-         outcome regardless of the relaxation trajectory (the same
-         convention `make_mcmm_proposer` uses).
-
-    Topology preservation is approximate, not strict: GOAT freezes
-    bonds and ring sp² angles via constraints during the uphill push,
-    while we let MMFF relax them. For peptide cyclisations and N-Me
-    bonds this is fine — MMFF's bond-stretch term has a steep gradient
-    that pulls covalent bonds back to equilibrium even from large
-    kicks. For sigma_kick_a beyond ~0.3 Å, expect occasional MMFF
-    non-convergence or bond-breaking; tune accordingly.
-
-    Params:
-        mol: Chem.Mol : reference molecule with explicit Hs. Topology
-            captured at factory build time; do not mutate structurally.
-        hardware_opts : nvmolkit hardware options for batched MMFF
-            (only used when `mmff_backend='gpu'`).
-        calc : MACECalculator from `get_mace_calc()`.
-        sigma_kick_a: float : Gaussian standard deviation in Å applied
-            independently to every atom-coordinate (default 0.1 Å).
-            Treat as the move-magnitude analogue of DBT's
-            `drive_sigma_rad`.
-        score_chunk_size: int : MACE per-batch forward pass cap
-            (default 500).
-        mmff_backend: str : 'gpu' (nvmolkit batched CUDA) or 'cpu'
-            (RDKit serial).
-        seed: int : RNG seed for the per-step kicks.
-    Returns:
-        callable : `batch_propose_fn(coords_list) -> list[tuple]`
-            matching the `ParallelMCMMDriver` /
-            `ReplicaExchangeMCMMDriver` proposer contract.
-    Raises:
-        ValueError: on unknown `mmff_backend`.
-    """
-    if mmff_backend not in ("gpu", "cpu"):
-        raise ValueError(
-            f"unknown mmff_backend {mmff_backend!r}; expected 'gpu' or 'cpu'"
-        )
-
-    n_atoms = mol.GetNumAtoms()
-    atomic_nums = [a.GetAtomicNum() for a in mol.GetAtoms()]
-    template_mol = Chem.Mol(mol)
-    template_mol.RemoveAllConformers()
-
-    rng = np.random.default_rng(seed)
-    stats = {"n_proposed": 0, "n_relax_failures": 0, "n_relax_successes": 0}
-
-    def batch_propose_fn(coords_list):
-        from confsweeper import _mace_batch_energies
-
-        n_walkers = len(coords_list)
-        if n_walkers == 0:
-            return []
-        stats["n_proposed"] += n_walkers
-
-        # Stage 1: per-walker isotropic Gaussian kick.
-        kicked_coords: list = []
-        for coords in coords_list:
-            coords_np = coords.detach().cpu().numpy().astype(np.float64)
-            kick = rng.normal(0.0, sigma_kick_a, size=coords_np.shape)
-            kicked_coords.append(coords_np + kick)
-
-        # Stage 2: batched MMFF on a fresh throwaway mol.
-        throwaway = Chem.Mol(template_mol)
-        for kc in kicked_coords:
-            conf = Chem.Conformer(n_atoms)
-            for a_idx in range(n_atoms):
-                x, y, z = kc[a_idx]
-                conf.SetAtomPosition(a_idx, (float(x), float(y), float(z)))
-            throwaway.AddConformer(conf, assignId=True)
-
-        if mmff_backend == "gpu":
-            from nvmolkit.mmffOptimization import MMFFOptimizeMoleculesConfs
-
-            MMFFOptimizeMoleculesConfs([throwaway], hardwareOptions=hardware_opts)
-        else:
-            from rdkit.Chem import AllChem as _AllChem
-
-            for cid in [c.GetId() for c in throwaway.GetConformers()]:
-                _AllChem.MMFFOptimizeMolecule(throwaway, confId=cid)
-
-        # Stage 3: batched MACE scoring, chunked.
-        post_mmff_conf_ids = [c.GetId() for c in throwaway.GetConformers()]
-        energies: list = []
-        for start in range(0, len(post_mmff_conf_ids), score_chunk_size):
-            chunk_ids = post_mmff_conf_ids[start : start + score_chunk_size]
-            ase_mols = [
-                ase.Atoms(
-                    positions=throwaway.GetConformer(cid).GetPositions(),
-                    numbers=atomic_nums,
-                )
-                for cid in chunk_ids
-            ]
-            energies.extend(_mace_batch_energies(calc, ase_mols))
-
-        # Stage 4: assemble per-walker proposals. det_j=1.0 for the
-        # symmetric Gaussian kick. Walkers whose post-MMFF energy is
-        # non-finite (MMFF blow-up — rare but possible at large
-        # sigma_kick_a) get success=False so the driver rejects.
-        proposals: list = []
-        for slot, cid in enumerate(post_mmff_conf_ids):
-            new_coords = torch.tensor(
-                throwaway.GetConformer(cid).GetPositions(), dtype=torch.float64
-            )
-            e = float(energies[slot])
-            if not np.isfinite(e):
-                stats["n_relax_failures"] += 1
-                proposals.append((coords_list[slot], 0.0, 0.0, False))
-            else:
-                stats["n_relax_successes"] += 1
-                proposals.append((new_coords, e, 1.0, True))
-        return proposals
-
-    batch_propose_fn.stats = stats
-    return batch_propose_fn
-
-
-# ---------------------------------------------------------------------------
-# Composite proposer — randomly route walkers to one of several sub-proposers
-# ---------------------------------------------------------------------------
-
-
-def make_composite_proposer(
-    proposers,
-    weights=None,
-    seed: int = 0,
-):
-    """
-    Build a `batch_propose_fn` that routes each walker to one of N
-    sub-proposers per step, sampled by `weights`. Lets DBT and
-    Cartesian-kick (or future move types) coexist in one MCMM run.
-
-    Routing happens at the walker level, not the step level — so a
-    single REMD step can have some walkers proposing DBT moves and
-    others proposing Cartesian kicks. Walkers are partitioned by
-    chosen proposer, each sub-proposer is invoked on its subset
-    (preserving its internal batching), and results are reassembled
-    in walker order.
-
-    Each sub-proposer's `.stats` dict is preserved on the composite
-    return value as `.stats[i]` so callers can inspect per-proposer
-    diagnostics. The composite itself does not aggregate counters.
-
-    Params:
-        proposers: list[callable] : sub-proposers, each matching the
-            `batch_propose_fn(coords_list) -> list[tuple]` contract.
-        weights: list[float] | None : sampling weight per proposer.
-            Default None means uniform. Normalised internally.
-        seed: int : routing RNG seed; deterministic across replicate
-            runs.
-    Returns:
-        callable : `batch_propose_fn(coords_list) -> list[tuple]`.
-            Carries `.stats = [p.stats for p in proposers]` (list,
-            indexed by proposer position).
-    Raises:
-        ValueError: empty `proposers`, weight/proposer count mismatch,
-            or any non-positive weight.
-    """
-    if not proposers:
-        raise ValueError("proposers must be non-empty")
-    if weights is None:
-        weights_arr = np.full(len(proposers), 1.0 / len(proposers))
-    else:
-        if len(weights) != len(proposers):
-            raise ValueError(
-                f"weights ({len(weights)}) must match proposers ({len(proposers)})"
-            )
-        if any(w <= 0 for w in weights):
-            raise ValueError(f"weights must be positive, got {weights}")
-        w_arr = np.asarray(weights, dtype=np.float64)
-        weights_arr = w_arr / w_arr.sum()
-
-    rng = np.random.default_rng(seed)
-
-    def batch_propose_fn(coords_list):
-        n = len(coords_list)
-        if n == 0:
-            return []
-        choices = rng.choice(len(proposers), size=n, p=weights_arr)
-
-        results: list = [None] * n
-        for p_idx, propose_fn in enumerate(proposers):
-            walker_idx = [w for w in range(n) if choices[w] == p_idx]
-            if not walker_idx:
-                continue
-            sub_coords = [coords_list[w] for w in walker_idx]
-            sub_results = propose_fn(sub_coords)
-            if len(sub_results) != len(sub_coords):
-                raise RuntimeError(
-                    f"proposer {p_idx} returned {len(sub_results)} results for "
-                    f"{len(sub_coords)} walkers"
-                )
-            for w, r in zip(walker_idx, sub_results):
-                results[w] = r
-        return results
-
-    batch_propose_fn.stats = [p.stats for p in proposers if hasattr(p, "stats")]
-    return batch_propose_fn
+from proposers import (  # noqa: E402
+    _backbone_atom_set,
+    _compute_window_downstream_sets,
+    _side_chain_group,
+    make_cartesian_kick_proposer,
+    make_composite_proposer,
+    make_mcmm_proposer,
+)
